+++ Sources/ViewModel/MessageThread/SBUMessageThreadViewModel.swift
@@ -301,7 +301,11 @@ open class SBUMessageThreadViewModel: SBUBaseChannelViewModel {
             if error.code == ChatError.nonAuthorized.rawValue {
                 self.delegate?.baseChannelViewModel(self, shouldDismissForChannel: nil)
             } else {
-                self.delegate?.didReceiveError(error, isBlocker: true)
+                if SendbirdChat.isLocalCachingEnabled {
+                    return true
+                } else {
+                    self.delegate?.didReceiveError(error, isBlocker: true)
+                }
             }
             return false
         }